Make EAF Browser history more flexible
- Add clear_history mapping to C-M-q to clear history
- Add eaf-browser-store-history variable for user to customize
- Move history functions to app/browser/buffer.py instead
