+++ core/browser.py
@@ -51,7 +51,6 @@ class BrowserView(QWebEngineView):
         self.cookie_store.cookieAdded.connect(self.cookie_storage.add_cookie)
 
         self.selectionChanged.connect(self.select_text_change)
-        self.titleChanged.connect(self.record_history)
 
         self.load_cookie()
 
@@ -62,18 +61,6 @@ class BrowserView(QWebEngineView):
 
         with open(os.path.join(os.path.dirname(__file__), "js", "goto_marker.js"), "r") as f:
             self.goto_marker_raw = f.read()
-
-    def record_history(self, title):
-        self.history_log_file_path = os.path.join(self.config_dir, "browser", "history", "log.txt")
-        touch(self.history_log_file_path)
-        with open(self.history_log_file_path, "a") as f:
-            filtered_url = self.filter_url(self.url().toString())
-            if title == self.url().toString():
-                f.write(filtered_url + " " + filtered_url + "\n")
-            else:
-                f.write(title + " " + filtered_url + "\n")
-
-        self.remove_duplicate_history()
 
     def filter_url(self, url):
         parsed = urlparse(url)
@@ -95,17 +82,6 @@ class BrowserView(QWebEngineView):
             urlencode(filtered, doseq=True), # query string
             parsed.fragment
         ])
-
-    def remove_duplicate_history(self):
-        lines_seen = set() # holds lines already seen
-        with open(self.history_log_file_path, "r") as f:
-            lines = f.readlines()
-
-        with open(self.history_log_file_path, "w") as f:
-            for line in lines:
-                if line not in lines_seen: # not a duplicate
-                    f.write(line)
-                    lines_seen.add(line)
 
     def _search_text(self, text, is_backward = False):
         if self.search_term != text:
@@ -415,9 +391,6 @@ class BrowserBuffer(Buffer):
     def get_key_event_widgets(self):
         # We need send key event to QWebEngineView's focusProxy widget, not QWebEngineView.
         return [self.buffer_widget.focusProxy()]
-
-    def get_history_log_file(self):
-        return self.buffer_widget.history_log_file_path
 
     def scroll(self, scroll_direction, scroll_type):
         if scroll_type == "page":
